refactor: move position_velocity.rs into ephemeris
