+++ src/utils/mod.rs
@@ -12,7 +12,6 @@ pub mod eop_provider;
 pub mod leap_seconds;
 pub mod math_utils;
 pub mod naif_ids;
-pub mod position_velocity;
 pub mod time_utils;
 pub mod to_skycoord;
 pub mod ut1_provider;
